Finalize removal of final decompositions and restructure the repo a bit
