+++ tests/test_optimization.py
@@ -20,17 +20,19 @@ import cirq
 import pytest
 from cirq import ops, optimizers
 
-from cirq_iqm.adonis import Adonis
-from cirq_iqm.iqm_device import (DropRZBeforeMeasurement,
-                                 MergeOneParameterGroupGates)
+from cirq_iqm.optimizers import (DropRZBeforeMeasurement,
+                                 MergeOneParameterGroupGates, simplify_circuit)
 
 TOLERANCE = 1e-10  # numerical tolerance
 
 
 @pytest.fixture(scope='module')
-def adonis():
-    """Adonis device fixture."""
-    return Adonis()
+def qubits():
+    return [
+        cirq.NamedQubit('Alice'),
+        cirq.NamedQubit('Bob'),
+        cirq.NamedQubit('Charlie')
+    ]
 
 
 class TestGateOptimization:
@@ -41,10 +43,10 @@ class TestGateOptimization:
         ops.ISwapPowGate,
     ])
     @pytest.mark.parametrize('a, b', [(0, 0.3), (-0.5, 0.5), (1.0, 2.0), (0.1, -4.1)])
-    def test_gate_merging(self, family, a, b):
+    def test_gate_merging(self, family, a, b, qubits):
         """Merging one-parameter group gates."""
 
-        q0, q1 = cirq.NamedQubit('q0'), cirq.NamedQubit('q1')
+        q0, q1 = qubits[:2]
 
         c = cirq.Circuit()
         c.append([
@@ -83,10 +85,10 @@ class TestGateOptimization:
             marks=pytest.mark.xfail(strict=True, reason='Implementation missing in Cirq.')
         ),
     ])
-    def test_eject_z(self, family, ex):
+    def test_eject_z(self, family, ex, qubits):
         """Commuting z rotations towards the end of the circuit."""
 
-        q0, q1 = cirq.NamedQubit('q0'), cirq.NamedQubit('q1')
+        q0, q1 = qubits[:2]
 
         c = cirq.Circuit()
         c.append([
@@ -107,52 +109,52 @@ class TestGateOptimization:
 class TestSimplifyCircuit:
 
     @pytest.mark.xfail(strict=True, reason='ZZPowGate does not yet implement the _phase_by_ protocol in Cirq.')
-    def test_simplify_circuit_eject_z(self, adonis):
+    def test_simplify_circuit_eject_z(self, qubits):
 
-        q0, q1 = adonis.qubits[:2]
+        q0, q1 = qubits[:2]
         c = cirq.Circuit()
         c.append([
             cirq.ZPowGate(exponent=0.55)(q0),
             cirq.ZZPowGate(exponent=0.1)(q0, q1),
             cirq.MeasurementGate(2)(q0, q1),
         ])
-        new = adonis.simplify_circuit(c)
+        new = simplify_circuit(c)
 
         # TODO ZPowGate should have been ejected and dropped
         assert len(new) == 2
 
-    def test_simplify_circuit_merge_one_parameter_gates(self, adonis):
+    def test_simplify_circuit_merge_one_parameter_gates(self, qubits):
 
-        q0, q1 = adonis.qubits[:2]
+        q0, q1 = qubits[:2]
         c = cirq.Circuit()
         c.append([
             cirq.ZZPowGate(exponent=0.3)(q0, q1),
             cirq.ZZPowGate(exponent=0.1)(q0, q1),
         ])
-        new = adonis.simplify_circuit(c)
+        new = simplify_circuit(c)
 
         # ZZPowGates have been merged
         assert len(new) == 1
 
-    def test_simplify_circuit_drop_rz_before_measurement(self, adonis):
+    def test_simplify_circuit_drop_rz_before_measurement(self, qubits):
 
-        q0, q1 = adonis.qubits[:2]
+        q0, q1 = qubits[:2]
         c = cirq.Circuit()
         c.append([
             cirq.ZPowGate(exponent=0.1)(q0),
             cirq.ZPowGate(exponent=0.2)(q1),
             cirq.MeasurementGate(1, key='measurement')(q0),
         ])
-        new = adonis.simplify_circuit(c, use_final_decomposition=False)
+        new = simplify_circuit(c)
 
         # the ZPowGate preceding the measurement has been dropped
         assert len(new) == 2
         assert isinstance(new[0].operations[0].gate, cirq.ZPowGate)
         assert new[0].operations[0].qubits == (q1,)
 
-    def test_drop_rz_before_measurement(self, adonis):
+    def test_drop_rz_before_measurement(self, qubits):
 
-        q0, q1, q2 = adonis.qubits[:3]
+        q0, q1, q2 = qubits[:3]
         c = cirq.Circuit()
         c.append([
             (cirq.X ** 0.4)(q0),
@@ -173,9 +175,9 @@ class TestSimplifyCircuit:
         assert isinstance(op.gate, cirq.ZPowGate)
         assert op.qubits == (q2,)
 
-    def test_drop_rz_before_measurement_drop_final(self, adonis):
+    def test_drop_rz_before_measurement_drop_final(self, qubits):
 
-        q0, q1, q2 = adonis.qubits[:3]
+        q0, q1, q2 = qubits[:3]
         c = cirq.Circuit()
         c.append([
             (cirq.X ** 0.4)(q0),
@@ -193,16 +195,16 @@ class TestSimplifyCircuit:
         assert isinstance(op.gate, cirq.XPowGate)
         assert op.qubits == (q0,)
 
-    def test_simplify_circuit_merge_one_qubit_gates(self, adonis):
+    def test_simplify_circuit_merge_one_qubit_gates(self, qubits):
 
-        q0 = adonis.qubits[0]
+        q0 = qubits[0]
         c = cirq.Circuit()
         c.append([
             cirq.XPowGate(exponent=0.1)(q0),
             cirq.YPowGate(exponent=0.2)(q0),
             cirq.ZPowGate(exponent=0.3)(q0),
         ])
-        new = adonis.simplify_circuit(c, use_final_decomposition=False)
+        new = simplify_circuit(c)
 
         # the one-qubit gates have been merged
         assert len(new) == 2
